+++ src/App.tsx
@@ -4,7 +4,7 @@ import { Routes, Route } from 'react-router-dom'
 import Home from '@/pages/home'
 import Level from '@/pages/level'
 import Result from '@/pages/result'
-import Exercice from '@/pages/exercice'
+import Exercice from '@/pages/exercise'
 import Layout from './components/layout/layout'
 import TheEnd from './pages/end'
 import ResultCompleted from './pages/completed'
