+++ src/components/code-generator.tsx
@@ -2,7 +2,7 @@ import saveCodeToDatabase from '@/services/service-code'
 import generateCode from '@/utils'
 import authenticateCode from '@/services/service-auth'
 import useAuthStore from '@/store/tracking/tracker-auth'
-import useLevelStore from '@/store/store-level'
+import useLevelStore from '@/store/level-progress'
 import useProgressStore from '@/store/tracking/tracker-progress'
 import useTotalQuizStore from '@/store/tracking/tracker-total-answer'
 import { useRef, useEffect, useState } from 'react'
