+++ src/components/has-code.tsx
@@ -2,7 +2,7 @@ import { useRef, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import authenticateCode from '@/services/service-auth'
 import useAuthStore from '@/store/tracking/tracker-auth'
-import useLevelStore from '@/store/store-level'
+import useLevelStore from '@/store/level-progress'
 import useProgressStore from '@/store/tracking/tracker-progress'
 import useTotalQuizStore from '@/store/tracking/tracker-total-answer'
 import { API_URL } from '@/config/api'
