+++ src/pages/exercise.tsx
@@ -1,17 +1,17 @@
 import Title from '@/components/title'
-import Progressbar from '@/components/progressbar'
-import ButtonHelp from '@/components/buttonhelp'
-import ProgressNumber from '@/components/progressnumber'
+import Progressbar from '@/components/progress-bar'
+import ButtonHelp from '@/components/button-help'
+import ProgressNumber from '@/components/progress-number'
 import CardAnswer from '@/components/card-answer'
-import CustomButton from '@/components/buttonlong'
+import CustomButton from '@/components/button-long'
 import useProgressStore from '@/store/tracking/tracker-progress'
 import useExerciseStore from '@/store/data/exercise'
-import useLevelStore from '@/store/store-level'
+import useLevelStore from '@/store/level-progress'
 import Feedback from '@/components/feedback'
 import useQuizStore from '@/store/tracking/tracker-answer'
 import { useState } from 'react'
 import useFeedbackStore from '@/store/data/feedback'
-import useAnswerStore from '@/store/store-choice-answer'
+import useAnswerStore from '@/store/selected-answer'
 import { useNavigate } from 'react-router-dom'
 import useAuthStore from '@/store/tracking/tracker-auth'
